+++ view/tws_administration_view/lib/view/articles/trucks/trucks_article.dart
@@ -1,6 +1,5 @@
 import 'package:csm_view/csm_view.dart';
 import 'package:flutter/material.dart';
-import 'package:http/http.dart';
 import 'package:tws_administration_view/core/extension/datetime.dart';
 import 'package:tws_administration_view/core/router/twsa_routes.dart';
 import 'package:tws_administration_view/data/services/sources.dart';