+++ view/tws_administration_view/lib/view/widgets/tws_article_creation/tws_article_creation_item_state.dart
@@ -23,4 +23,9 @@ final class TWSArticleCreatorItemState<TModel> extends CSMStateBase {
     _valid = newInvalid;
     effect();
   }
+
+  void updateFactory(TModel Function() useFactory){
+    _model = useFactory();
+    effect();
+  }
 }